add app layout
