+++ src/frontend/src/assets/scss/layout/header.scss
@@ -53,6 +53,9 @@
 }
 
 .header__user {
+  display: flex;
+  align-items: center;
+
   a {
     display: block;
